Create day_3_idea
idea 1st :- cattle monitoring

problem statement :-  there are IoT agriculture sensors that can be attached to the animals on a farm to monitor their health and log performance. Livestock tracking and monitoring help collect data on stock health, well-being, and physical location.
And for this health monitoring , we are able to prevent humans from virus and infections which are spread by animals.

such sensors can identify sick animals so that farmers can separate them from the herd and avoid contamination. Using drones for real-time cattle tracking also helps farmers reduce staffing expenses. This works similarly to IoT devices for petcare.

idea :- 2nd :  Paying guest and landlord verification for Rent.

Problem statement ;- We want to design such a platform, where the paying guest and landlord verify itself with their document, and it is finally verified by police.

so it is reduces the manual work and paper work, where paper may will loss, but this platform will overcome this problem,
and also prevent the fraud which can be happen with our hardcopies of  documents..